# chore: Improve ApprovalModal performance