… app/components/Approvals/ApprovalModal/__snapshots__/ApprovalModal.test.tsx.snap
… app/components/Snaps/SnapUIRenderer/__snapshots__/SnapUIRenderer.test.tsx.snap